refactor: cleaning function
